pyproject.toml
[tool.poetry]
name = "test-finance-task"
version = "0.1.0"
description = ""
authors = ["nickpashkevich <micolas9441@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.5.0"
drf-spectacular = "^0.27.2"
black = "^24.10.0"
django-filter = "^24.3"
watchdog = "^5.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
python-json-logger = "^2.0.7"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]